Moved code templates and help content into a new sub-package
